files renamed additional. Imports checked
